+++ packages/orq-rc/RELEASES.md
@@ -7599,3 +7599,13 @@ Based on:
 - [python v4.9.0-rc.27] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.27] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.27 - packages/orq-rc
+
+## 2026-05-06 13:03:16
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.28] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.28] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.28 - packages/orq-rc